Rename /ui/main/src/app/business/common in /ui/main/src/app/utils (#7790)
