+++ ui/main/src/app/business/view/card/card-body.view.ts
@@ -9,7 +9,7 @@
 import {Card} from '@ofServices/cards/model/Card';
 import {ShowAcknowledgmentFooterEnum} from '@ofServices/processes/model/Processes';
 import {UserWithPerimeters} from '@ofServices/users/model/UserWithPerimeters';
-import {Utilities} from 'app/business/common/utilities';
+import {Utilities} from '../../../utils/utilities';
 import {ProcessesService} from '@ofServices/processes/ProcessesService';
 import {ConfigService} from 'app/services/config/ConfigService';
 import {UserPermissionsService} from '@ofServices/userPermissions/UserPermissionsService';
