+++ ui/main/src/app/business/view/dashboard/dashboard.view.spec.ts
@@ -16,7 +16,7 @@ import {OpfabEventStreamService} from '@ofServices/events/OpfabEventStreamServic
 import {getOneLightCard, setProcessConfiguration, setUserPerimeter} from '@tests/helpers';
 import {firstValueFrom, skip} from 'rxjs';
 import {Severity} from '@ofModel/light-card.model';
-import {Utilities} from 'app/business/common/utilities';
+import {Utilities} from '../../../utils/utilities';
 import {FilterType} from '@ofModel/feed-filter.model';
 import {FilteredLightCardsStore} from '../../../store/lightcards/lightcards-feed-filter-store';
 import {OpfabStore} from '../../../store/opfabStore';
