+++ ui/main/src/app/business/view/processmonitoring/processmonitoring.view.ts
@@ -10,7 +10,7 @@
 import {ProcessesService} from '@ofServices/processes/ProcessesService';
 import {ProcessGroups, ProcessToMonitor, StatePerProcessToMonitor} from './processmonitoringPage';
 import {UsersService} from '@ofServices/users/UsersService';
-import {Utilities} from 'app/business/common/utilities';
+import {Utilities} from '../../../utils/utilities';
 import {Process} from '@ofServices/processes/model/Processes';
 import {PermissionEnum} from '@ofServices/groups/model/PermissionEnum';
 import {add, endOfWeek, format, parse, startOfWeek, sub} from 'date-fns';
