+++ ui/main/src/app/components/archives/archives.component.ts
@@ -24,9 +24,9 @@ import {ConfigService} from 'app/services/config/ConfigService';
 import {NgbModal, NgbModalOptions, NgbModalRef, NgbPopover, NgbPagination} from '@ng-bootstrap/ng-bootstrap';
 import {LightCard} from '@ofModel/light-card.model';
 import {Page} from '@ofModel/page.model';
-import {ExcelExport} from 'app/business/common/excel-export';
+import {ExcelExport} from '../../utils/excel-export';
 import {UserPreferencesService} from '@ofServices/userPreferences/UserPreferencesService';
-import {Utilities} from 'app/business/common/utilities';
+import {Utilities} from '../../utils/utilities';
 import {Card, CardWithChildCards} from '@ofServices/cards/model/Card';
 import {ArchivesLoggingFiltersComponent} from '../share/archives-logging-filters/archives-logging-filters.component';
 import {DisplayContext} from '@ofModel/template.model';
