+++ ui/main/src/app/components/feed/components/pinned-cards/pinned-cards.component.ts
@@ -11,7 +11,7 @@ import {OnInit, Component, OnDestroy} from '@angular/core';
 import {ProcessesService} from '@ofServices/processes/ProcessesService';
 import {Subject, takeUntil, timer} from 'rxjs';
 import {LightCard} from '@ofModel/light-card.model';
-import {Utilities} from 'app/business/common/utilities';
+import {Utilities} from '../../../../utils/utilities';
 import {OpfabStore} from '../../../../store/opfabStore';
 import {NgFor, NgIf, LowerCasePipe} from '@angular/common';
 import {NgbPopover} from '@ng-bootstrap/ng-bootstrap';
