+++ ui/main/src/app/components/processmonitoring/processmonitoring.component.ts
@@ -24,7 +24,7 @@ import {ConfigService} from 'app/services/config/ConfigService';
 import {DateTimeFormatterService} from 'app/services/dateTimeFormatter/DateTimeFormatterService';
 import {LightCard} from '@ofModel/light-card.model';
 import {Page} from '@ofModel/page.model';
-import {ExcelExport} from 'app/business/common/excel-export';
+import {ExcelExport} from '../../utils/excel-export';
 import {EntitiesService} from '@ofServices/entities/EntitiesService';
 import {NgbModal, NgbModalOptions, NgbModalRef} from '@ng-bootstrap/ng-bootstrap';
 import {CardsFilter} from '@ofServices/cards/model/CardsFilter';
