+++ ui/main/src/app/components/share/countdown/countdown.component.ts
@@ -10,7 +10,7 @@
 import {Component, Input, OnChanges, OnDestroy, OnInit} from '@angular/core';
 import {TranslateService, TranslateModule} from '@ngx-translate/core';
 import {ConfigService} from 'app/services/config/ConfigService';
-import {CountDown} from '../../../business/common/countdown/countdown';
+import {CountDown} from '../../../utils/countdown/countdown';
 import {NgIf} from '@angular/common';
 
 @Component({
