+++ ui/main/src/app/services/admin/CrudUtils.ts
@@ -7,7 +7,7 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {Utilities} from 'app/business/common/utilities';
+import {Utilities} from '../../utils/utilities';
 import {MultiSelectOption} from 'app/components/share/multi-select/model/MultiSelect';
 import {EntitiesService} from '../entities/EntitiesService';
 import {GroupsService} from '../groups/GroupsService';
