+++ ui/main/src/app/utils/countdown/countdown.spec.ts
