+++ ui/main/src/app/utils/jsontoarray/json-to-array.ts
