+++ ui/main/src/app/utils/utilities.ts
